+++ pkg/exams/R/search_files.R
@@ -51,4 +51,6 @@ match_exams_call <- function(which = 1L, deparse = TRUE) {
 
 match_exams_device <- function() .exams_get_internal("xweave_device")
 
+match_exams_markup <- function() .exams_get_internal("xweave_markup")
+
 match_exams_iteration <- function() .exams_get_internal("xexams_iteration")